changing project name
