+++ go.mod
@@ -1,3 +1,3 @@
-module pic_type
+module space-memes
 
-go 1.21.5
+go 1.21.0
